docs(language): update documentation to reflect updated language
Problem: The documentation was out of date given updated language
features such as expression based syntax. The document was also
incorrectly formatted in markdown style.

Solution: This updates the documentation so that it is up to date with
how the language has developed over time.

Notes: The language has seemed to evolve as I have been developing it. I
should keep this in mind for future implementations to update the docs
when new features are implemented.