feat(lexer): parse arrays for existing types
Problem: The lexer did not support the following syntax:

`string[]`

Solution: This implements this feature and serializes it as it's own
token.

Notes: This implementation ended up refactoring the logic to parse data
types into it's own function (which is for the better).

The implementation also makes zero assumptions about input, which is why
trim was employed everywhere. This should hopefully mean the input is
more sanitized in the future.

There is a problem with this implementation in that it doesn't support
2+d arrays, which should be supported in the future.