feat(lexer): implement support for comments
Problem: The lexer doesn't support single line and multipline comments.

Solution: Implemented a this with a `take_if` and `take_while` methods
to make this as declarative as possible.

Notes: This was hard, and I'm sure that there are some bugs in this
logic somewhere due to the amount of mutability. A lot more testing is
required.