fix(lexer): implement array token
Problem: The current implementation joins the type and it's array into 1
token. This is problematic because the lexing is doing too much for one
task.

Solution: This creates an `Array` token which consists of `[]`. This
will then be handed off to the parsing step to decide what is being
declared.

Notes: The implementation felt wrong before implementing this change.
The lexer was trying to encode type information which is not it's job
and that was making things complicated. This is a much better solution.