feat(lexer/parser): divide math operation into it's own type
Problem: The current implementation lumped math operations with the rest
of the symbols. This leads to the need for | statements in matches in the parsing,
stage which don't allow for type casting.

Solution: To fix this, we make the math symbol a type in itself, which
allows us to treat this type specifically.

Notes: This might need to be applied to the assign variants too.