feat(parser): implement if else expressions
Problem: The language didn't support if else statements being encoded
into the AST.

Solution: This implements this alongside making parsing changes. Most
notable is the use of a parsing state, which acts as a immutable
structure representing the "state" of the parse.

This means that a token stream is no longer passed around, enabling
referencing to previous states.

Notes: I'm glad this has finally worked, but the implementation of this
is getting too messy now. This parsing logic is getting too difficult to
reason about and trace. I want to work on a new solution that is more
declarative and scalable.