feat(parser): convert to immutable style
Problem: If else expressions showed that the current design philosophies
of the parser weren't working. There was simply too much complexity and
tracing with the token stream. There needed to be a better solution and
some code cleanup.

Solution: This commit addresses this problem by doing the following:

1. Converting the iterative loops to a recursive solution. This makes
   reasoning about the individual loops much easier.
2. "Immutability" by default. Strictly, the loops mentioned above aren't
immutable, but they take the same ideas of preprending to a vector and
then calling a function recursively.
3. Declarative pattern matching. This is helped also by recursion, as
   there is only one layer of complexity to think about (the function)
   instead of a full loop.

Notes: This is still open to change, but I think this immutable,
declarative style is a lot easier to read and debug. Especially as the
complexity of the parser grows and more features are added.