feat(parser): implement for loops
Problem: The parser didn't support the for loop syntax.

Solution: This adds support for for loops.

Notes: This led to some realizations about determining token context.

When we are starting out parsing, we try to figure out the context.
Therefore, we have little to no information on what the language
construct might be.

For instance:

Token::Identifier x

By itself, x can mean an identifier reference, a function declaration, a
function call, and other operations. Therefore, this needs to be
handled in the `on_identifier` functionality.

However, if we already have the context:

Token::Identifier x
Token::Symbol ParenOpen
Token::Symbol ParenClose
Token::Symbol BraceOpen
Token::Symbol BraceClose

At the `ParenOpen` token, we know this is a function declaration. And so
all future identifiers should be handled in `on_function` or some
function call in `on_function`.

If this is not done, and things are handled in `on_identifier`, context
is missing, and we can't figure out what a token might mean given the
next token.