feat(parser): implement constants
Problem: The parser didn't support assigning constants.

Solution: This adds a new field to Operation::Assignment to handle this
case.

Notes: Initially this started with assigning this to single, but
then I realised this should be in the assignment logic instead, since
this will apply to assignment operations, not values. This was
much more refactorable.

Tests cover don't cover identifier case but this can be added in the
future. Literals are tested.