fix(parser, language): split operations into separate types
Problem: The operation types were grouped together even though
operations represent different contexts.

Solution: Operations are broken out into OneHandedOperation and
TwoHandedOperation.

Notes: This came when developing for the runtime. It would be extremely
verbose to check all variants when the types can narrow down what is
going on in the current context.